w3c-reports/12092014/gtk/runned-tests/html/semantics/embedded-content/media-elements/track/track-element/cors/009-expected.txt

Blocked access to external URL http://www1.localhost:8800/html/semantics/embedded-content/media-elements/track/track-element/cors/support/set-cookie.html#f9e0e191-2420-4e57-8ac4-9b1ad1194765
CONSOLE MESSAGE: Cross-origin text track load denied by Cross-Origin Resource Sharing policy.
Blocked access to external URL http://www1.localhost:8800/html/semantics/embedded-content/media-elements/track/track-element/cors/support/cors-tester.py?delete-cookie&id=f9e0e191-2420-4e57-8ac4-9b1ad1194765
Test id: f9e0e191-2420-4e57-8ac4-9b1ad1194765



PASS track CORS: No CORS, not same-origin, no headers 
